Fix warnings from static analysis tools
1.INVALID ASSIGNMENT:
  a. phDnldNfc.cc: phDnldNfc_LoadFW()
  The size of variables equated are different.
  "fileSize" is 4 (unsigned int) and "fTellFileSize" is 8(long)

  b. phDnldNfc_Internal.cc: phDnldNfc_CreateFramePld()
  Explicit type casting is done to avoid ambiguity.

  c. phNxpNciHal.cc: phNxpNciHal_core_initialized()
  The size of variables equated are different.
  "timeout" is 4 (unsigned int) and "retLen*1000" is
  8 (long)

  d. phNxpNciHal.cc: phNxpNciHal_getVendorConfig()
  The size of variables equated are different.
  "config.maxIsoDepTranscieveLength" is 4 (unsigned int)
	   and "num" is 8 (unsigned long)

  e. phOsalNfc_Timer.cc: phOsalNfc_DeferredCall()
  "dwIndex" datatype change from uint32_t to unsigned long

2.DO_NOT_USE_INT_IN_FOR_LOOP
  a. phNxpConfig.cpp: GetNxpNumValue()
   unsigned int is changed to size_t datatype

Bug: 142626701
Test: Nfc-On/Off, Tag read, HCE, VTS

+++ halimpl/hal/phNxpNciHal.cc
@@ -1667,7 +1667,7 @@ int phNxpNciHal_core_initialized(uint16_t core_init_rsp_params_len, uint8_t* p_c
       // Check the permissible range [0 - 60]
       if (0 <= retlen && retlen <= 60) {
         if (0 < retlen) {
-          unsigned int timeout = retlen * 1000;
+          unsigned int timeout = (uint32_t)retlen * 1000;
           unsigned int timeoutHx = 0x0000;
 
           char tmpbuffer[10] = {0};
@@ -2258,33 +2258,32 @@ void phNxpNciHal_getVendorConfig(android::hardware::nfc::V1_1::NfcConfig& config
 
   config.nfaPollBailOutMode = true;
   if (GetNxpNumValue(NAME_ISO_DEP_MAX_TRANSCEIVE, &num, sizeof(num))) {
-    config.maxIsoDepTransceiveLength = num;
+    config.maxIsoDepTransceiveLength = (uint32_t)num;
   }
   if (GetNxpNumValue(NAME_DEFAULT_OFFHOST_ROUTE, &num, sizeof(num))) {
-    config.defaultOffHostRoute = num;
+    config.defaultOffHostRoute = (uint8_t)num;
   }
   if (GetNxpNumValue(NAME_DEFAULT_NFCF_ROUTE, &num, sizeof(num))) {
-    config.defaultOffHostRouteFelica = num;
+    config.defaultOffHostRouteFelica = (uint8_t)num;
   }
   if (GetNxpNumValue(NAME_DEFAULT_SYS_CODE_ROUTE, &num, sizeof(num))) {
-    config.defaultSystemCodeRoute = num;
+    config.defaultSystemCodeRoute = (uint8_t)num;
   }
   if (GetNxpNumValue(NAME_DEFAULT_SYS_CODE_PWR_STATE, &num, sizeof(num))) {
-    config.defaultSystemCodePowerState = num;
+    config.defaultSystemCodePowerState = (uint8_t)num;
   }
   if (GetNxpNumValue(NAME_DEFAULT_ROUTE, &num, sizeof(num))) {
-    config.defaultRoute = num;
+    config.defaultRoute = (uint8_t)num;
   }
   if (GetNxpByteArrayValue(NAME_DEVICE_HOST_WHITE_LIST, (char*)buffer.data(), buffer.size(), &retlen)) {
     config.hostWhitelist.resize(retlen);
-    for(int i=0; i<retlen; i++)
-      config.hostWhitelist[i] = buffer[i];
+    for(long i=0; i<retlen; i++) config.hostWhitelist[i] = buffer[i];
   }
   if (GetNxpNumValue(NAME_OFF_HOST_ESE_PIPE_ID, &num, sizeof(num))) {
-    config.offHostESEPipeId = num;
+    config.offHostESEPipeId = (uint8_t)num;
   }
   if (GetNxpNumValue(NAME_OFF_HOST_SIM_PIPE_ID, &num, sizeof(num))) {
-    config.offHostSIMPipeId = num;
+    config.offHostSIMPipeId = (uint8_t)num;
   }
   if ((GetNxpByteArrayValue(NAME_NFA_PROPRIETARY_CFG, (char*)buffer.data(), buffer.size(), &retlen))
          && (retlen == 9)) {
